requirements.txt
streamlit
sentence-transformers
scikit-learn
openai
python-dotenv
PyPDF2
python-docx

